+++ lib/contentful/deleted_entry.rb
@@ -1,10 +1,7 @@
-require_relative 'resource'
+require_relative 'base_resource'
 
 module Contentful
   # Resource class for deleted entries
   # https://www.contentful.com/developers/documentation/content-delivery-api/http/#sync-item-types
-  class DeletedEntry
-    include Contentful::Resource
-    include Contentful::Resource::SystemProperties
-  end
+  class DeletedEntry < BaseResource; end
 end