+++ lib/contentful/locale.rb
@@ -1,13 +1,13 @@
-require_relative 'resource'
-
 module Contentful
   # A Locale definition as included in Space
   # Read more about Localization at https://www.contentful.com/developers/documentation/content-delivery-api/#i18n
   class Locale
-    include Contentful::Resource
+    attr_reader :code, :name, :default
 
-    property :code, :string
-    property :name, :string
-    property :default, :boolean
+    def initialize(json)
+      @code = json.fetch('code', nil)
+      @name = json.fetch('name', nil)
+      @default = json.fetch('default', false)
+    end
   end
 end