+++ lib/contentful/location.rb
@@ -1,12 +1,14 @@
-require_relative 'resource'
-
 module Contentful
   # Location Field Type
   # You can directly query for them: https://www.contentful.com/developers/documentation/content-delivery-api/#search-filter-geo
   class Location
-    include Contentful::Resource
+    attr_reader :lat, :lon
+    alias latitude lat
+    alias longitude lon
 
-    property :lat, :float
-    property :lon, :float
+    def initialize(json)
+      @lat = json.fetch('lat', nil)
+      @lon = json.fetch('lon', nil)
+    end
   end
 end